Fix incosistent wasm/pc time
Now time is measured from UNIX_EPOCH on both platforms.

+++ native/sapp-wasm/src/rand.rs
@@ -1,5 +1,5 @@
 pub const RAND_MAX: u32 = 2147483647;
 extern "C" {
     pub fn rand() -> ::std::os::raw::c_int;
-    pub fn time() -> f32;
+    pub fn time() -> f64;
 }
