+++ src/lib.rs
@@ -42,7 +42,7 @@ pub mod date {
 
     #[cfg(target_arch = "wasm32")]
     pub fn now() -> f64 {
-        unsafe { sapp::time() as f64 }
+        unsafe { sapp::time() }
     }
 }
